pyproject.toml
[tool.poetry]
name = "balance-api"
version = "0.1.0"
description = ""
authors = ["Juliano Vidal <juliano.vidal@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.6.5"
backoff = "^1.11.1"
connexion = "^2.9.0"
Flask-Compress = "^1.10.1"
Flask-Cors = "^3.0.10"
Flask-SQLAlchemy = "^2.5.1"
meza = "^0.42.5"
psycopg2-binary = "^2.9.1"
python-jose = "^3.3.0"
SQLAlchemy = "^1.4.22"
flake8 = "^3.9.2"
mypy = "^0.910"
six = "^1.16.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
